# docs/openapi.yaml
# Здесь будет OpenAPI описание всех доступных эндпоинтов и моделей данных.
openapi: 3.0.3
info:
  title: Bank REST API
  version: 1.0.0
  description: Backend API for managing bank cards with JWT security
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login and obtain JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: JWT issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JwtResponse'

  /api/cards:
    get:
      tags: [User Cards]
      summary: Get current user's cards (paginated)
      parameters:
        - in: query
          name: page
          schema: { type: integer, minimum: 0 }
        - in: query
          name: size
          schema: { type: integer, minimum: 1 }
        - in: query
          name: sort
          schema: { type: string }
      responses:
        '200':
          description: Page of cards
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/CardResponse'
                  totalElements: { type: integer }
                  totalPages: { type: integer }
                  number: { type: integer }
                  size: { type: integer }

  /api/cards/{id}:
    get:
      tags: [User Cards]
      summary: Get a specific own card by id
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Card
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'

  /api/cards/balance/{id}:
    get:
      tags: [User Cards]
      summary: Get balance of own card
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Balance
          content:
            application/json:
              schema:
                type: number
                format: double

  /api/cards/transfer:
    post:
      tags: [User Cards]
      summary: Transfer between own cards
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferRequest'
      responses:
        '200': { description: Transferred }

  /api/admin/cards:
    post:
      tags: [Admin Cards]
      summary: Create card for user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCardRequest'
      responses:
        '200':
          description: Created card
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
    get:
      tags: [Admin Cards]
      summary: Get all cards (paginated)
      parameters:
        - in: query
          name: page
          schema: { type: integer, minimum: 0 }
        - in: query
          name: size
          schema: { type: integer, minimum: 1 }
        - in: query
          name: sort
          schema: { type: string }
      responses:
        '200':
          description: Page of cards
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/CardResponse'
                  totalElements: { type: integer }
                  totalPages: { type: integer }
                  number: { type: integer }
                  size: { type: integer }

  /api/admin/cards/{id}/activate:
    patch:
      tags: [Admin Cards]
      summary: Activate card
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Activated card
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'

  /api/admin/cards/{id}/block:
    patch:
      tags: [Admin Cards]
      summary: Block card
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Blocked card
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'

  /api/admin/cards/{id}:
    delete:
      tags: [Admin Cards]
      summary: Delete card
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '204': { description: Deleted }

  /api/admin/users:
    post:
      tags: [Admin Users]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '200':
          description: Created user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
    get:
      tags: [Admin Users]
      summary: Get users
      responses:
        '200':
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'

  /api/admin/users/{id}:
    get:
      tags: [Admin Users]
      summary: Get user by id
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: User
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
    delete:
      tags: [Admin Users]
      summary: Delete user by id
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '204': { description: Deleted }

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Role:
      type: string
      enum: [ADMIN, USER]
    CardStatus:
      type: string
      enum: [ACTIVE, BLOCKED, EXPIRED]
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string }
    JwtResponse:
      type: object
      properties:
        token: { type: string }
        username: { type: string }
        role: { type: string }
    CreateUserRequest:
      type: object
      required: [username, password, role]
      properties:
        username: { type: string }
        password: { type: string }
        role:
          $ref: '#/components/schemas/Role'
    UserResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        username: { type: string }
        role:
          $ref: '#/components/schemas/Role'
        enabled: { type: boolean }
    CreateCardRequest:
      type: object
      required: [userId, cardNumber, expirationDate]
      properties:
        userId: { type: integer, format: int64 }
        cardNumber: { type: string }
        expirationDate: { type: string, format: date }
    CardResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        maskedCardNumber: { type: string, example: '**** **** **** 1234' }
        expirationDate: { type: string, format: date }
        status:
          $ref: '#/components/schemas/CardStatus'
        balance: { type: number, format: double }
    TransferRequest:
      type: object
      required: [fromCardId, toCardId, amount]
      properties:
        fromCardId: { type: integer, format: int64 }
        toCardId: { type: integer, format: int64 }
        amount: { type: number, format: double, minimum: 0.01 }